--- ShinEdit/Page_1/__init__.py
@@ -1,4 +1,0 @@
-from . import debtordata
-from . import kpidata
-from . import productdata
-from . import salesdata
