+++ ShinEdit/Web_Page/__init__.py
@@ -1,0 +1,3 @@
+from . import CRUD_page
+from . import dashboard_page
+from . import login_page
